Migrate another team endpoint to new pattern
