+++ server/service/endpoint_teams.go
@@ -2,57 +2,10 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/go-kit/kit/endpoint"
 )
-
-////////////////////////////////////////////////////////////////////////////////
-// Modify Team Agent Options
-////////////////////////////////////////////////////////////////////////////////
-
-type modifyTeamAgentOptionsRequest struct {
-	ID      uint
-	options json.RawMessage
-}
-
-func makeModifyTeamAgentOptionsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(modifyTeamAgentOptionsRequest)
-		team, err := svc.ModifyTeamAgentOptions(ctx, req.ID, req.options)
-		if err != nil {
-			return teamResponse{Err: err}, nil
-		}
-
-		return teamResponse{Team: team}, err
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Delete Team
-////////////////////////////////////////////////////////////////////////////////
-
-type deleteTeamRequest struct {
-	ID uint `json:"id"`
-}
-
-type deleteTeamResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r deleteTeamResponse) error() error { return r.Err }
-
-func makeDeleteTeamEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(deleteTeamRequest)
-		err := svc.DeleteTeam(ctx, req.ID)
-		if err != nil {
-			return deleteTeamResponse{Err: err}, nil
-		}
-		return deleteTeamResponse{}, nil
-	}
-}
 
 ////////////////////////////////////////////////////////////////////////////////
 // List Team Users
