+++ server/service/service_teams.go
@@ -2,18 +2,9 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/fleetdm/fleet/v4/server/fleet"
 )
-
-func (svc *Service) ModifyTeamAgentOptions(ctx context.Context, id uint, options json.RawMessage) (*fleet.Team, error) {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
-	svc.authz.SkipAuthorization(ctx)
-
-	return nil, fleet.ErrMissingLicense
-}
 
 func (svc *Service) AddTeamUsers(ctx context.Context, teamID uint, users []fleet.TeamUser) (*fleet.Team, error) {
 	// skipauth: No authorization check needed due to implementation returning
@@ -45,14 +36,6 @@ func (svc *Service) ListAvailableTeamsForUser(ctx context.Context, user *fleet.U
 	svc.authz.SkipAuthorization(ctx)
 
 	return nil, fleet.ErrMissingLicense
-}
-
-func (svc *Service) DeleteTeam(ctx context.Context, tid uint) error {
-	// skipauth: No authorization check needed due to implementation returning
-	// only license error.
-	svc.authz.SkipAuthorization(ctx)
-
-	return fleet.ErrMissingLicense
 }
 
 func (svc *Service) TeamEnrollSecrets(ctx context.Context, teamID uint) ([]*fleet.EnrollSecret, error) {
